requirements.txt
py-cord[voice]
PyNaCl
aiodns
brotlipy
cchardet
python-dotenv
